mqtt-python-server/requirements.txt
paho-mqtt==1.6.1
mysql-connector-python==8.0.33
psutil==5.9.5
